chore(dsp-contract-negotiation-transform): refactor rebase
